requirements.txt
click==8.0.1
Flask==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
python-http-client==3.3.2
sendgrid==6.8.0
starkbank-ecdsa==1.1.1
Werkzeug==2.0.1
